add readin page
