+++ src/components/Navbar/navData.js
@@ -1,11 +1,11 @@
 export const navLinks = [
   {
-    label: 'Projects',
-    pathname: '/#project-1',
-  },
-  {
     label: 'Details',
     pathname: '/#details',
+  },
+  {
+    label: 'Projects',
+    pathname: '/#project-1',
   },
   {
     label: 'Articles',
