+++ src/pages/projects/readin/VolkiharKnight.module.css
@@ -1,0 +1,24 @@
+.textSection {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.logoContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0 80px;
+  width: 100%;
+  overflow: hidden;
+
+  & svg {
+    max-width: 400px;
+    height: auto;
+  }
+}
+
+.armor {
+  aspect-ratio: 1 / 2;
+  display: grid;
+}
