+++ src/pages/projects/readin/index.page.js
@@ -1,0 +1,1 @@
+export { VolkiharKnight as default } from './VolkiharKnight';
