--- src/pages/projects/volkihar-knight/index.page.js
@@ -1,1 +1,0 @@
-export { VolkiharKnight as default } from './VolkiharKnight';
